smart-trash-bin/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
emails
jinja2
